Add vote unit tests
Added unit tests for vote, then rewrote the original functions to use raw queries.
I'm going to be converting to raw queries just for better control and understanding of the db.

Next tests will need to include for sql injection, and I'll need to update the queries to use
parameterized queries to prevent it.