Add test cases for inserting new questions to the DB
I had a tough time figuring out why my questionId column kept displaying
null, even though it was being passed valid number. Turns out the FK relationship
will cause the FK to be NULL'd if the parent row were to be deleted.

I don't know if this is a feature of MySQL, or the Sequelize.